% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activatorGet.R
\name{activatorGet}
\alias{activatorGet}
\title{Return list of data.frames of activator activities from POTA API}
\usage{
activatorGet(call, url = "https://api.pota.app/profile/")
}
\arguments{
\item{call}{ham radio call to be queried}

\item{url}{is default to (https://api.pota.app/profile)}
}
\value{
A list of data.frames for activator
}
\description{
Return list of data.frames of activator activities from POTA API
}
\examples{
call <- "ke4mkg" ;
activatorGet(call)
}
