% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkHist.R
\name{getParkHist}
\alias{getParkHist}
\title{Return data.frame of activations from POTA API}
\usage{
getParkHist(park, url = "https://api.pota.app/park/activations/")
}
\arguments{
\item{park}{park identifier to be queried}

\item{url}{is default to (https://api.pota.app/profile)}
}
\value{
A list of data.frames for activator
}
\description{
Return data.frame of activations from POTA API
}
\examples{


park <- "US-2177" 
getParkHist(park)
}
