% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParks.R
\name{getParks}
\alias{getParks}
\title{Return data.frame of parks from POTA API}
\usage{
getParks(llat, llon, ulat, ulon, url = "https://api.pota.app/park/grids")
}
\arguments{
\item{llat}{Lower left latitude coordinate (decimal degrees)}

\item{llon}{Lower left longitude coordinate}

\item{ulat}{upper right latitude coordinate}

\item{ulon}{upper right longitude coordinate}

\item{url}{is default to (https://api.pota.app/park/grids)}
}
\value{
A data.frame of parks
}
\description{
Return data.frame of parks from POTA API
}
\examples{
library(dplyr)
library(sf)

x <- "EM83aw"
qth <- maidenhead("EM83aw") |>
st_point() |>
st_sfc(crs = 4326)

circleDist <- qth |> st_buffer(dist = 1E5, max_cells = 10000) 
bbox <- circleDist |>
st_bbox()
parks_sf <- getParks(bbox[2], bbox[1], bbox[4], bbox[3]) |>
rowwise() |>
mutate(longitude = geometry.coordinates[1],
         latitude = geometry.coordinates[2]) |>
           st_as_sf(coords = c("longitude", "latitude"), crs = 4326)
}
