% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maidenhead.R
\name{maidenhead}
\alias{maidenhead}
\title{Return Latitude and Longitude from 6 digit Maidenhead Grid}
\usage{
maidenhead(x)
}
\arguments{
\item{x}{A character vector with one element that is a valid Maidenhead Grid.}
}
\value{
A numeric vector of latitude and longitude.
}
\description{
Return Latitude and Longitude from 6 digit Maidenhead Grid
}
\examples{
x <- "EM83aw"
maidenhead(x)
}
