% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParksPath.R
\name{ParksPath}
\alias{ParksPath}
\title{Return data.frame of parks from POTA API}
\usage{
ParksPath(route, buffer = 5)
}
\arguments{
\item{route}{3 features read from kml path}

\item{buffer}{distance in kilometers}
}
\value{
A sf data.frame of parks
}
\description{
Return data.frame of parks from POTA API
}
\examples{

x <- sf::st_read(system.file("extdata", "directions.kml", package = "potaplan"))
ParksPath(x)

}
