% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParksRad.R
\name{ParksRad}
\alias{ParksRad}
\title{Return data.frame of parks from POTA API}
\usage{
ParksRad(x, buffer = 100)
}
\arguments{
\item{x}{valid 6 digit Maidenhead}

\item{buffer}{distance in kilometers}
}
\value{
A sf data.frame of parks
}
\description{
Return data.frame of parks from POTA API
}
\examples{
ParksRad("EM83aw")
}
