% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potaplanner_radius.R
\name{potaplanner_radius}
\alias{potaplanner_radius}
\title{Creates a leaflet map showing if parks in radius of a Maidenhead have been worked by a given hamcall.}
\usage{
potaplanner_radius(hamcall = "KE4MKG", grid = "EM83aw", distance = 100)
}
\arguments{
\item{hamcall}{is hamcall to check}

\item{grid}{is maidenhead grid (6 digits)}

\item{distance}{is distance in kilometers to check parks}
}
\value{
A leaflet map
}
\description{
Creates a leaflet map showing if parks in radius of a Maidenhead have been worked by a given hamcall.
}
\examples{


hamcall <- "KE4MKG" 
maidenhead <- "EM83aw"
radius <- 100

potaplanner_radius(hamcall, maidenhead, radius)
}
